requirements.txt
streamlit
python-dotenv
faiss-cpu
sentence-transformers
pdfplumber
python-pptx
python-docx
pandas
openai
requests
PyMuPDF
